Added maxRetries parameter to extractSubtrees task
I neglected to include this key => value pair in the initial implementation.  With larger and larger input files the workflow may fail on Terra with an 'OutOfMemory' type error.  By including this parameter to the runtime block in the WDL code the user can then select the "Retry with more Memory" option before launching the workflow.  Without adding this parameter, selecting the "Retry with more Memory" option will have no effect.